refactor(config): derive memory search types from schema
