+++ src/config/types.memory.ts
@@ -1,9 +1,8 @@
-import type { MemoryExtraPath } from "../memory-host-sdk/host/types.js";
 /**
  * Memory config types shared by core context-engine paths and memory host/plugin runtimes.
  * Builtin memory stays core-owned.
  */
-import type { SecretInput } from "./types.secrets.js";
+import type { MemorySearchConfigInput } from "./zod-schema.memory-search.js";
 
 export type { MemoryExtraPath } from "../memory-host-sdk/host/types.js";
 
@@ -17,86 +16,12 @@ export type MemoryConfig = {
   search?: MemorySearchConfig;
 };
 
-export type MemorySearchConfig = {
-  /** Enable vector memory search (default: true). */
-  enabled?: boolean;
-  /** Use relevant context from this agent's other private conversations. */
-  rememberAcrossConversations?: boolean;
-  /** Sources to index and search (default: ["memory"]). */
-  sources?: Array<"memory" | "sessions">;
-  /** Extra paths to include in memory search, optionally filtered by a glob. */
-  extraPaths?: MemoryExtraPath[];
-  /** Optional multimodal file indexing for selected extra paths. */
-  multimodal?: {
-    /** Enable image/audio embeddings from extraPaths. */
-    enabled?: boolean;
-    /** Which non-text file types to index. */
-    modalities?: Array<"image" | "audio" | "all">;
-    /** Max bytes allowed per multimodal file before it is skipped. */
-    maxFileBytes?: number;
-  };
-  /** Experimental session transcript indexing. */
-  experimental?: {
-    sessionMemory?: boolean;
-  };
-  /** Memory embedding provider adapter id. */
-  provider?: string;
-  remote?: {
-    baseUrl?: string;
-    apiKey?: SecretInput;
-    headers?: Record<string, string>;
-    batch?: {
-      /** Enable batch API for embedding indexing (OpenAI/Gemini; default: true). */
-      enabled?: boolean;
-    };
-  };
-  /** Fallback memory embedding provider adapter id when embeddings fail. */
-  fallback?: string;
-  /** Embedding model id (remote) or alias (local). */
-  model?: string;
-  /** Optional provider-specific embedding input_type for query and document requests. */
-  inputType?: string;
-  /** Optional provider-specific embedding input_type for query-time memory search. */
-  queryInputType?: string;
-  /** Optional provider-specific embedding input_type for document/index embeddings. */
-  documentInputType?: string;
-  /**
-   * Provider-specific output vector dimensions. Gemini supports 128 to 3072.
-   * Google recommends 768, 1536, or 3072 dimensions.
-   */
-  outputDimensionality?: number;
-  /** Local embedding settings for the managed llama.cpp server. */
-  local?: {
-    /** GGUF model path or hf: URI. */
-    modelPath?: string;
-  };
-  /** Index storage configuration. */
-  store?: {
-    fts?: {
-      /** FTS5 tokenizer (default: "unicode61"). Use "trigram" for CJK text support. */
-      tokenizer?: "unicode61" | "trigram";
-    };
-    vector?: {
-      /** Enable the sqlite-vec semantic index (default: true). */
-      enabled?: boolean;
-      /** Optional override path to sqlite-vec extension (.dylib/.so/.dll). */
-      extensionPath?: string;
-    };
+export type MemorySearchConfig = Omit<MemorySearchConfigInput, "store"> & {
+  /** Preserve legacy embedding-cache authoring accepted by Doctor migrations. */
+  store?: NonNullable<MemorySearchConfigInput["store"]> & {
     cache?: {
-      /** Enable embedding cache (default: true). */
       enabled?: boolean;
-      /** Optional max cache entries per provider/model. */
       maxEntries?: number;
     };
-  };
-  /** Query behavior. */
-  query?: {
-    maxResults?: number;
-    minScore?: number;
-  };
-  /** Index cache behavior. */
-  cache?: {
-    /** Cache chunk embeddings in SQLite (default: true). */
-    enabled?: boolean;
   };
 };
